// src/app/pages/license/api/validation-result.interface.ts
export enum ValidationStep {
    STEP_CONTENT_LIBRARY_EXISTS = 'SMC_LICENSE.INSTALLATION.CONTENTLIB_EXISTS',
    STEP_CONTENT_LIBRARY_ACCESS = 'SMC_LICENSE.INSTALLATION.CONTENTLIB_ACCESS',
    STEP_QLIK_LICENSE_ACCESS    = 'SMC_LICENSE.INSTALLATION.QLIK_LICENSE_ACCESS',
    STEP_QLIK_LICENSE_VALID     = 'SMC_LICENSE.INSTALLATION.QLIK_LICENSE_VALID'
}

export interface ILicenseValidationResult {

    isValid: boolean;

    errors: string[];
}